[cxx-interop] Cache empty lifetime dependencies for methods
If we could not infer lifetime dependencies for a C++ method, let's cache the empty result. This prevents the compiler from trying to evaluate this request using Swift Sema's algorithm.

This change is meant to fix the following issue:

With AddressableParameters enabled, instantiations of `std::optional<T>` lack the conformance to `protocol CxxOptional`, despite the synthesized conformance being added correctly to each instantiation.

This happens because when importing `operator*()` of each instantiation, ClangImporter spins up a LifetimeDependenceInfoRequest for the incompletely imported C++ type. In particular, the synthesized protocol conformances aren't yet added at that point, since the conformance logic relies on type's members already being imported. The LifetimeDependenceInfoRequest triggers a ConformanceTable to be built and cached for the instantiation. Later, by the time the conformance synthesis logic runs, the ConformanceTable for the instantiation had already been populated, and won't be updated to add the new conformance.