Add unsafe labels for strict memory safety
This Swift standard library builds with strict memory safety on,
which means that the compiler issues a warning whenever unsafe C
capabilities such as raw pointers are used. To silence the warning
the 'unsafe' annotation must be used to acknowledge the lack of
safety.

This annotation should be added to the FreeBSD implementation of
the Mutex type in the Synchronization module.