Update SwiftCompilerSources AllocBoxToStack to handle mark_dep.
Handle the easy case in which the promoted box is the base operand of a
mark_dep. This requires nothing more than setting the base operand to the new
stack address.

Required for move-only checking. Otherwise, we get an incredibly bizarre
compiler error: noncopyable 'foo' cannot be consumed when captured by an
escaping closure

The legacy pass was updated first for 6.2.