[CMake] Ensure incremental builds invokes the Runtimes build system...
...when using `--enable-new-runtime-build`

When building changes incrementally the runtime build system is not
invoked -- the configure and build steps generated by `ExternalProject_Add`
are effectively triggered only if the `CMAKE_ARGS` argument changes.
We should instead delegate to the the Runtime build system to figure out
if it needs to regenerate the `build.ninja` file and/or trigger any
rebuild -- this can be done by passing `BUILD_ALWAYS 1` to
`ExternalProject_Add`.

Addresses rdar://155559208