[cxx-interop] Fix assert failure due to invalid location
We hit an assert failure do to creating a range where the begin was a
valid source location while the end range was an invalid one.
Unfortunately, I was not able to create a minimal reproducer but I was
able to verify that this fix works because I was able to reproduce this
issue while compiling Swift with a patched version of Clang.

rdar://155379027