Runtimes: correct the Android overlay install more precisely
The install location of the APINotes is separate from the rest of the
overlay content. These are installed unlabelled by the OS (which is a
potential issue if there are conflicts with other OSes!). The swift
swift-driver properly locates the APINotes it appears, but the C++
swift-driver does not. Install the APINotes to a location that both will
function with.