Runtimes: export the SwiftOverlay targets
This allows wiring up dependencies across the projects. In particular,
this is needed to get the generated VFS mapping in the Android build for
the explicit dependencies to work. The PCM emission does not receive the
required mapping and thus is unable to load the modulemaps for the build
of the PCM.