Runtimes: reworking Distributed build to work on Windows
The primary change required here is to remove the `LINKER_LANGUAGE`
property on the target with policy CMP0157 set to `OLD`. Take the
opportunity to re-order the flag handling to match the other runtime
library CMakeLists making it easier to diff them. `-library-level api`
is implicitly handled by the `emit_swift_interface` call, so remove
that. The other runtimes use a centralised `ExperimentalFeatures` which
would add in the unsafe support. While in the area, add in a dependency
on `SwiftOverlay` to wire up the `builtin_float` module dependency as
well.