Build-Script: FreeBSD Package Preset
Adding a preset for building the FreeBSD toolchain and runtime package.
This is very similar to the Linux package build, but the main difference
is the lack of WASM build. The WASM runtime build system uses GNU
extensions in its makefiles that are not compatible with the make or
gmake on FreeBSD and therefore cannot be built there.

(cherry picked from commit dbb3498236280b1e65a65dd836ed37f5805ab828)