Switch to a linear-space variant of Myers diffing
This changes the implementation for `Collection.difference(from:)`
to use a linear-space complexity variation of the same Myers
algorithm. The new version is similar in execution time to the
existing one, but should alleviate memory pressure when diffing
collections where the number of differences approaches the size of
the collection. While the new algorithm returns a set of changes
that is the same size as the previous version, the specific changes
are not guaranteed to be the same.