Serialization: Reject swiftmodules built with a different C++ interop mode
Intro an option to reject swiftmodules built in a different C++ interop
mode. This rejects non-library-evolution swiftmodules that are
incompabtible with a fatal error, and forces rebuilding
library-evolution enabled modules from the swiftinterface. This applies
to both local and distributed modules.

This option is behind the env var
SWIFT_ENABLE_SWIFTMODULE_PER_CXX_INTEROP for now.

Allow for modules under the resource dir to be be rebuilt from
swiftinterface on a C++ interop mismatch.