WIP: Implement swift-inspect dump-concurrency for Windows
The main pieces remaining are to provide an API in swift_Concurrency.dll
to provide the address of the Swift task pointer thread_local variable,
or its fully computed offset, since doing so requires access to the
address. In this prototype I hardcode '8' for now.

Secondly, iterating the heap results in WaitForSingleObject hitting a
timeout when reading from the shared memory. I haven't investigated
this, but Saleem said it worked when he originally implemented it.