Do not use the -file-compilation-dir path for pch
With de425be15b5fba898cfbe3a1a434302101e14c27 we made it so that if
a -file-compilation-dir path was passed in, the skeleton CU's
DW_AT_comp_dir was set to that path. However, we also set the PCH path
to the -file-compilation-dir path which caused a regression.

This patch fixes that issue.

(cherry picked from commit e4a06e895d20449bfe0d4e6f1aed1193642eed9b)