[Test][Concurrency] Fix reflect_task test to avoid a race condition.
This isn't a nice fix, because I can't see a good way to fix this problem
robustly.  The issue is that the Concurrency runtime may mutate the AsyncTask
structure from a thread other than the one that is reflecting the current
task, so looking through at the child tasks is inherently dangerous and
may cause the test to crash at random, even though it will *often* work.

To make this a lot less likely, add a .25s delay before reflecting the
parent task.

rdar://150478597